requirements_project.txt
# requirements_project.txt
# German Lecture Slide OCR with Meta-Learning
# Hardware: RTX 4060 Ti 16GB, CUDA 11.8+

# ─────────────────────────────────────────
# Core Deep Learning
# ─────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0

# ─────────────────────────────────────────
# Computer Vision
# ─────────────────────────────────────────
opencv-python>=4.9.0
opencv-contrib-python>=4.9.0
Pillow>=10.1.0
albumentations>=1.3.0
timm>=0.9.0

# ─────────────────────────────────────────
# Detection (YOLOv8 baseline)
# ─────────────────────────────────────────
ultralytics>=8.0.0

# ─────────────────────────────────────────
# Meta-Learning
# ─────────────────────────────────────────
learn2learn>=0.1.7

# ─────────────────────────────────────────
# German NLP
# ─────────────────────────────────────────
spacy>=3.7.0
pyspellchecker>=0.7.3

# ─────────────────────────────────────────
# Math Processing
# ─────────────────────────────────────────
sympy>=1.12
sacrebleu>=2.3.0

# ─────────────────────────────────────────
# Data Processing
# ─────────────────────────────────────────
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ─────────────────────────────────────────
# Visualization
# ─────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
tensorboard>=2.15.0

# ─────────────────────────────────────────
# Utilities
# ─────────────────────────────────────────
tqdm>=4.66.0
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0

# ─────────────────────────────────────────
# Evaluation
# ─────────────────────────────────────────
evaluate>=0.4.0
jiwer>=3.0.0

# ─────────────────────────────────────────
# Document Processing
# ─────────────────────────────────────────
lxml>=4.9.1
beautifulsoup4>=4.12.0

# ─────────────────────────────────────────
# TAMER dependencies (already cloned)
# ─────────────────────────────────────────
einops>=0.7.0
editdistance>=0.6.0
pytorch-lightning>=1.8.0,<2.0.0
jsonargparse[signatures]>=4.0.0
typer>=0.4.1
